Style fix and improvements (WIP)
