+++ lib/frontend/src/actions/index.js
@@ -79,6 +79,10 @@ export const updateRepoLabels = missingAndOverlyLabels => ({
   missingAndOverlyLabels,
 })
 
+export const emptyForm = () => ({
+  type: 'EMPTY_FORM',
+})
+
 export const stateToParams = (state, target) => {
   let users
   switch (target) {
@@ -370,5 +374,14 @@ export const deleteLabels = () => async (dispatch, getState) => {
   }
   if (allErrors !== []) {
     dispatch(setErrorRepo(keyIndex(allErrors, 1)))
+  }
+}
+
+export const goBack = history => dispatch => {
+  dispatch(emptyForm())
+  if (history.length === 1) {
+    history.push('/')
+  } else {
+    history.go(-1)
   }
 }
