+++ lib/frontend/src/actions/issueForm.js
@@ -1,13 +1,9 @@
-import { changeRepository, setError } from './index'
+import { changeRepository, goBack, setError } from './index'
 import { sortMilestones } from '../utils'
 
 export const setErrorCreate = error => ({
   type: 'SET_ERROR_CREATE',
   error,
-})
-
-export const emptyForm = () => ({
-  type: 'EMPTY_FORM',
 })
 
 export const updateRolesSelect = rolesSelect => ({
@@ -177,13 +173,4 @@ export const submitIssue = (event, history) => async (dispatch, getState) => {
   return dispatch(
     setErrorCreate(`${response.status}: ${response.statusText} ${jsonMessage}`)
   )
-}
-
-export const goBack = history => dispatch => {
-  dispatch(emptyForm())
-  if (history.length === 1) {
-    history.push('/')
-  } else {
-    history.go(-1)
-  }
 }
