+++ lib/frontend/src/components/Circle.jsx
@@ -3,9 +3,9 @@ import './Circle.sass'
 import { stringify } from 'query-string'
 import React from 'react'
 import { Helmet } from 'react-helmet'
-import { Link } from 'react-router-dom'
+import { Link, withRouter } from 'react-router-dom'
 
-import { block, connect } from '../utils'
+import { editCircle, goBack } from '../actions'
 import {
   deleteCircle,
   toggleAccountExpanded,
@@ -13,7 +13,7 @@ import {
   togglePurposeExpanded,
   updateCircle,
 } from '../actions/circle'
-import { editCircle } from '../actions'
+import { block, connect } from '../utils'
 import Loading from './Loading'
 import MarkdownEditor from './MarkdownEditor'
 
@@ -32,6 +32,7 @@ function Circle({
   circles,
   editClick,
   error,
+  history,
   loading,
   meetingsTypes,
   users,
@@ -39,6 +40,7 @@ function Circle({
   onClickDomain,
   onClickPurpose,
   onEdit,
+  onGoBack,
 }) {
   return (
     <section className={b()}>
@@ -62,273 +64,282 @@ function Circle({
           )}
         </article>
       )}
-      {circle.circle_name && (
-        <div>
-          <h1>{circle.circle_name}</h1>
-          {circle.parent_circle_name && (
-            <span>
-              {circle.parent_circle_name
-                ? `(sous-cercle de "${circle.parent_circle_name}")`
-                : ''}
-            </span>
-          )}
+      <article className={b('circle')}>
+        {circle.circle_name && (
+          <div>
+            <h1>{circle.circle_name}</h1>
+            {circle.parent_circle_name && (
+              <span>
+                {circle.parent_circle_name
+                  ? `(sous-cercle de "${circle.parent_circle_name}")`
+                  : ''}
+              </span>
+            )}
 
-          {loading && <Loading />}
-          {circle.is_in_edition ? (
-            <form
-              onSubmit={e => {
-                e.preventDefault()
-                onEdit(circle.circle_id, e)
-              }}
-            >
-              <h1>Edit {circle.circle_name} circle :</h1>
-              <label>
-                Name :
-                <input
-                  name="circle_name"
-                  defaultValue={circle.circle_name}
-                  required
-                />
-              </label>
-              <br />
-              <label>
-                Parent :
-                <select name="parent_circle_id" required>
-                  {circle.parent_circle_id === null ? (
-                    <option value=""> Aucun </option>
-                  ) : (
-                    <option defaultValue={circle.parent_circle_id}>
-                      {circle.parent_circle_name}
-                    </option>
-                  )}
-                  {circles
-                    .filter(cercle => cercle.circle_id !== circle.circle_id)
-                    .map(cercle => (
-                      <option key={cercle.circle_id} value={cercle.circle_id}>
-                        {cercle.circle_name}
+            {loading && <Loading />}
+            {circle.is_in_edition ? (
+              <form
+                onSubmit={e => {
+                  e.preventDefault()
+                  onEdit(circle.circle_id, e)
+                }}
+              >
+                <h1>Edit {circle.circle_name} circle :</h1>
+                <label>
+                  Name :
+                  <input
+                    name="circle_name"
+                    defaultValue={circle.circle_name}
+                    required
+                  />
+                </label>
+                <br />
+                <label>
+                  Parent :
+                  <select name="parent_circle_id" required>
+                    {circle.parent_circle_id === null ? (
+                      <option value=""> Aucun </option>
+                    ) : (
+                      <option defaultValue={circle.parent_circle_id}>
+                        {circle.parent_circle_name}
                       </option>
-                    ))}
-                </select>
-              </label>
-              <br />
-              <label>
-                Purpose :
-                <input
-                  name="circle_purpose"
-                  defaultValue={circle.circle_purpose}
-                  required
-                />
-              </label>
-              <br />
-              <label>
-                Domain :
-                <input
-                  name="circle_domain"
-                  defaultValue={circle.circle_domain}
-                  required
-                />
-              </label>
-              <br />
-              <label>
-                Accountabilities :
-                <MarkdownEditor />
-              </label>
-              <br />
-              <button type="submit">Edit</button>
-            </form>
-          ) : (
-            <div>
-              <article>
-                <h3>Purpose</h3>
-                <div onClick={() => onClickPurpose(circle.purpose_expanded)}>
-                  {circle.purpose_expanded ? (
-                    <p>{circle.circle_purpose}</p>
+                    )}
+                    {circles
+                      .filter(cercle => cercle.circle_id !== circle.circle_id)
+                      .map(cercle => (
+                        <option key={cercle.circle_id} value={cercle.circle_id}>
+                          {cercle.circle_name}
+                        </option>
+                      ))}
+                  </select>
+                </label>
+                <br />
+                <label>
+                  Purpose :
+                  <input
+                    name="circle_purpose"
+                    defaultValue={circle.circle_purpose}
+                    required
+                  />
+                </label>
+                <br />
+                <label>
+                  Domain :
+                  <input
+                    name="circle_domain"
+                    defaultValue={circle.circle_domain}
+                    required
+                  />
+                </label>
+                <br />
+                <label>
+                  Accountabilities :
+                  <MarkdownEditor />
+                </label>
+                <br />
+                <button type="submit">Edit</button>
+              </form>
+            ) : (
+              <div>
+                <article>
+                  <h3>Purpose</h3>
+                  <div onClick={() => onClickPurpose(circle.purpose_expanded)}>
+                    {circle.purpose_expanded ? (
+                      <p>{circle.circle_purpose}</p>
+                    ) : (
+                      <span>show purpose</span>
+                    )}
+                  </div>
+                  <h3>Domains</h3>
+                  <div onClick={() => onClickDomain(circle.domain_expanded)}>
+                    {circle.domain_expanded ? (
+                      <p>{circle.circle_domain}</p>
+                    ) : (
+                      <span>show domain</span>
+                    )}
+                  </div>
+                  <h3>Accountabilities</h3>
+                  <div
+                    onClick={() =>
+                      onClickAccount(circle.accountabilities_expanded)
+                    }
+                  >
+                    {circle.accountabilities_expanded ? (
+                      <ReactMarkdown source={circle.circle_accountabilities} />
+                    ) : (
+                      <span>show accountabilities</span>
+                    )}
+                  </div>
+                </article>
+                <article>
+                  <button type="submit" onClick={() => editClick()}>
+                    {circle.is_in_edition ? 'Cancel' : 'Update'}
+                  </button>
+                  {circle.roles && circle.roles.length > 0 ? (
+                    <span>
+                      <button type="submit" disabled>
+                        Delete
+                      </button>
+                      <br />
+                      <code>
+                        You cannot delete this circle, please first delete the
+                        roles.
+                      </code>
+                    </span>
                   ) : (
-                    <span>show purpose</span>
-                  )}
-                </div>
-                <h3>Domains</h3>
-                <div onClick={() => onClickDomain(circle.domain_expanded)}>
-                  {circle.domain_expanded ? (
-                    <p>{circle.circle_domain}</p>
-                  ) : (
-                    <span>show domain</span>
-                  )}
-                </div>
-                <h3>Accountabilities</h3>
-                <div
-                  onClick={() =>
-                    onClickAccount(circle.accountabilities_expanded)
-                  }
-                >
-                  {circle.accountabilities_expanded ? (
-                    <ReactMarkdown source={circle.circle_accountabilities} />
-                  ) : (
-                    <span>show accountabilities</span>
-                  )}
-                </div>
-              </article>
-              <article>
-                <button type="submit" onClick={() => editClick()}>
-                  {circle.is_in_edition ? 'Cancel' : 'Update'}
-                </button>
-                {circle.roles && circle.roles.length > 0 ? (
-                  <span>
-                    <button type="submit" disabled>
+                    <button
+                      type="submit"
+                      onClick={e => {
+                        e.preventDefault()
+                        btnClick(circle.circle_id)
+                      }}
+                    >
                       Delete
                     </button>
-                    <br />
-                    <code>
-                      You cannot delete this circle, please first delete the
-                      roles.
-                    </code>
-                  </span>
-                ) : (
-                  <button
-                    type="submit"
-                    onClick={e => {
-                      e.preventDefault()
-                      btnClick(circle.circle_id)
-                    }}
-                  >
-                    Delete
-                  </button>
-                )}
-              </article>
-              <article>
-                <h3>Meetings</h3>
-                {meetingsTypes.map(type => (
-                  <Link
-                    className={b('link')}
-                    key={type.type_id}
-                    to={{
-                      pathname: '/meetings',
-                      search: stringify({
-                        circle_id: circle.circle_id,
-                        meeting_name: type.type_name,
-                      }),
-                    }}
-                  >
-                    <button key={type.type_id} type="submit">
-                      {type.type_name}
-                    </button>
-                  </Link>
-                ))}
-              </article>
-              <article>
-                <h3>Rôles</h3>
-                {circle.roles && circle.roles.length > 0 ? (
-                  <ul>
-                    {circle.roles.map(role => (
-                      <li key={role.role_id} className={b('role')}>
-                        <span className={b('bullet')} />
-                        <Link
-                          to={{
-                            pathname: '/role',
-                            search: stringify({ id: role.role_id }),
-                          }}
-                        >
-                          {role.role_name}
-                        </Link>{' '}
-                        :{' '}
-                        <img
-                          key={role.user_id}
-                          className={b('avatar')}
-                          src={users
-                            .filter(user => getUserInfo(role.user_id, user))
-                            .map(user => user.avatar_url)
-                            .toString()}
-                          alt="avatar"
-                          title={users
+                  )}
+                </article>
+                <article>
+                  <h3>Meetings</h3>
+                  {meetingsTypes.map(type => (
+                    <Link
+                      className={b('link')}
+                      key={type.type_id}
+                      to={{
+                        pathname: '/meetings',
+                        search: stringify({
+                          circle_id: circle.circle_id,
+                          meeting_name: type.type_name,
+                        }),
+                      }}
+                    >
+                      <button key={type.type_id} type="submit">
+                        {type.type_name}
+                      </button>
+                    </Link>
+                  ))}
+                </article>
+                <article>
+                  <h3>Rôles</h3>
+                  {circle.roles && circle.roles.length > 0 ? (
+                    <ul>
+                      {circle.roles.map(role => (
+                        <li key={role.role_id} className={b('role')}>
+                          <span className={b('bullet')} />
+                          <Link
+                            to={{
+                              pathname: '/role',
+                              search: stringify({ id: role.role_id }),
+                            }}
+                          >
+                            {role.role_name}
+                          </Link>{' '}
+                          :{' '}
+                          <img
+                            key={role.user_id}
+                            className={b('avatar')}
+                            src={users
+                              .filter(user => getUserInfo(role.user_id, user))
+                              .map(user => user.avatar_url)
+                              .toString()}
+                            alt="avatar"
+                            title={users
+                              .filter(user => getUserInfo(role.user_id, user))
+                              .map(user => user.user_name)
+                              .toString()}
+                          />
+                          {'  '}
+                          {users
                             .filter(user => getUserInfo(role.user_id, user))
                             .map(user => user.user_name)
                             .toString()}
-                        />
-                        {'  '}
-                        {users
-                          .filter(user => getUserInfo(role.user_id, user))
-                          .map(user => user.user_name)
-                          .toString()}
-                      </li>
-                    ))}
-                  </ul>
-                ) : (
-                  <span>No roles defined</span>
-                )}
-                <Link
-                  to={{
-                    pathname: '/createrole',
-                    search: stringify({ circle_id: circle.circle_id }),
-                  }}
-                >
-                  <button type="submit">Add a Role</button>
-                </Link>
-              </article>
-            </div>
-          )}
-        </div>
-      )}
+                        </li>
+                      ))}
+                    </ul>
+                  ) : (
+                    <span>No roles defined</span>
+                  )}
+                  <Link
+                    to={{
+                      pathname: '/createrole',
+                      search: stringify({ circle_id: circle.circle_id }),
+                    }}
+                  >
+                    <button type="submit">Add a Role</button>
+                  </Link>
+                </article>
+              </div>
+            )}
+          </div>
+        )}
+        <br />
+        <button type="submit" onClick={() => onGoBack(history)}>
+          Back
+        </button>
+      </article>
     </section>
   )
 }
-export default connect(
-  state => ({
-    circle: state.circle.results,
-    circles: state.circles.results,
-    error: state.circle.error,
-    loading: state.circle.loading,
-    meetingsTypes: state.meetingsTypes.results,
-    users: state.users.results,
-  }),
-  dispatch => ({
-    btnClick: data => {
-      dispatch(deleteCircle(data))
-    },
-    editClick: () => {
-      dispatch(editCircle())
-    },
-    onClickAccount: circleAccount => {
-      dispatch(toggleAccountExpanded(circleAccount))
-    },
-    onClickDomain: circleDomain => {
-      dispatch(toggleDomainExpanded(circleDomain))
-    },
-    onClickPurpose: circlePurpose => {
-      dispatch(togglePurposeExpanded(circlePurpose))
-    },
-    onEdit: (id, e) => {
-      let formCircle = []
-      if (e.target.elements[1].value === '') {
-        formCircle = [
-          e.target.elements[0],
-          e.target.elements[2],
-          e.target.elements[3],
-          e.target.elements[4],
-          e.target.elements[5],
-        ].reduce(function(map, obj) {
-          if (obj.name === 'body') {
-            map.circle_accountabilities = obj.value
-          } else if (obj.name) {
-            map[obj.name] = obj.value
-          }
 
-          return map
-        }, {})
-      } else {
-        formCircle = [].slice
-          .call(e.target.elements)
-          .reduce(function(map, obj) {
-            if (obj.name) {
+export default withRouter(
+  connect(
+    state => ({
+      circle: state.circle.results,
+      circles: state.circles.results,
+      error: state.circle.error,
+      loading: state.circle.loading,
+      meetingsTypes: state.meetingsTypes.results,
+      users: state.users.results,
+    }),
+    dispatch => ({
+      btnClick: data => {
+        dispatch(deleteCircle(data))
+      },
+      editClick: () => {
+        dispatch(editCircle())
+      },
+      onClickAccount: circleAccount => {
+        dispatch(toggleAccountExpanded(circleAccount))
+      },
+      onClickDomain: circleDomain => {
+        dispatch(toggleDomainExpanded(circleDomain))
+      },
+      onClickPurpose: circlePurpose => {
+        dispatch(togglePurposeExpanded(circlePurpose))
+      },
+      onEdit: (id, e) => {
+        let formCircle = []
+        if (e.target.elements[1].value === '') {
+          formCircle = [
+            e.target.elements[0],
+            e.target.elements[2],
+            e.target.elements[3],
+            e.target.elements[4],
+            e.target.elements[5],
+          ].reduce(function(map, obj) {
+            if (obj.name === 'body') {
+              map.circle_accountabilities = obj.value
+            } else if (obj.name) {
               map[obj.name] = obj.value
             }
 
             return map
           }, {})
-      }
-      dispatch(updateCircle(id, formCircle))
-    },
-    onGoBack: () => {
-      history.go(-1)
-    },
-  })
-)(Circle)
+        } else {
+          formCircle = [].slice
+            .call(e.target.elements)
+            .reduce(function(map, obj) {
+              if (obj.name) {
+                map[obj.name] = obj.value
+              }
+
+              return map
+            }, {})
+        }
+        dispatch(updateCircle(id, formCircle))
+      },
+      onGoBack: history => {
+        dispatch(goBack(history))
+      },
+    })
+  )(Circle)
+)
