+++ lib/frontend/src/components/Createcircle.jsx
@@ -1,17 +1,23 @@
 import './Createcircle.sass'
 
 import React from 'react'
+import { Helmet } from 'react-helmet'
+import { withRouter } from 'react-router-dom'
 
-import { block, connect } from '../utils'
+import { goBack } from '../actions'
 import { createCircle } from '../actions/circle'
+import { block, connect } from '../utils'
 import Loading from './Loading'
 import MarkdownEditor from './MarkdownEditor'
 
 const b = block('Createcircle')
 
-function Createcircle({ loading, onSubmit, circles }) {
+function Createcircle({ circles, history, loading, onGoBack, onSubmit }) {
   return (
     <article className={b()}>
+      <Helmet>
+        <title>Gibolt - Create a circle</title>
+      </Helmet>
       {loading && <Loading />}
       <h2>Create a new circle :</h2>
       <form
@@ -55,30 +61,39 @@ function Createcircle({ loading, onSubmit, circles }) {
         </label>
         <br />
         <button type="submit"> Create circle</button>
+        <button type="submit" onClick={() => onGoBack(history)}>
+          Cancel
+        </button>
       </form>
     </article>
   )
 }
-export default connect(
-  state => ({
-    circles: state.circles.results,
-    loading: state.circles.loading,
-    error: state.circles.errors,
-  }),
-  dispatch => ({
-    onSubmit: e => {
-      const formCircle = [].slice
-        .call(e.target.elements)
-        .reduce(function(map, obj) {
-          if (obj.name === 'body') {
-            map.circle_accountabilities = obj.value
-          } else if (obj.name && obj.value) {
-            map[obj.name] = obj.value
-          }
 
-          return map
-        }, {})
-      dispatch(createCircle(formCircle))
-    },
-  })
-)(Createcircle)
+export default withRouter(
+  connect(
+    state => ({
+      circles: state.circles.results,
+      loading: state.circles.loading,
+      error: state.circles.errors,
+    }),
+    dispatch => ({
+      onGoBack: history => {
+        dispatch(goBack(history))
+      },
+      onSubmit: e => {
+        const formCircle = [].slice
+          .call(e.target.elements)
+          .reduce(function(map, obj) {
+            if (obj.name === 'body') {
+              map.circle_accountabilities = obj.value
+            } else if (obj.name && obj.value) {
+              map[obj.name] = obj.value
+            }
+
+            return map
+          }, {})
+        dispatch(createCircle(formCircle))
+      },
+    })
+  )(Createcircle)
+)
