+++ lib/frontend/src/components/Meeting.sass
@@ -16,3 +16,4 @@
     margin: 2em
     padding: 1em
     outline: none
+    width: 50em
